ODS-SERVER/requirements.txt
rembg
moviepy
ultralytics
django
opencv-python
numpy
pillow
